A-Packages: Organize into Packages
- Reorganized the classes into suitable java packages.
- Added Peer Review Suggestions.
